backend_server.py

from fastapi import FastAPI, File, UploadFile, Form, HTTPException
from fastapi.responses import JSONResponse
from pydantic import BaseModel
import uvicorn
import asyncio
import uuid
import json
from typing import Optional, Dict, Any
import numpy as np
from PIL import Image
import io
import requests
import os

app = FastAPI(title="MirrorWorld AI Backend", version="1.0.0")

# In-memory job storage (use Redis in production)
jobs: Dict[str, Dict[str, Any]] = {}

class JobStatus(BaseModel):
    job_id: str
    current_stage: str
    progress: float
    is_completed: bool
    assets: Optional[Dict[str, str]] = None
    error: Optional[str] = None

class ProcessingPipeline:
    """RenderNet + PIKE inspired AI pipeline for 3D avatar generation"""
    
    STAGES = [
        ("uploading", "Uploading to backend...", 0.1),
        ("segmentation", "MODNet person segmentation...", 0.2),
        ("depth_analysis", "MiDaS depth estimation...", 0.3),
        ("mesh_generation", "RenderNet 3D mesh creation...", 0.5),
        ("texture_mapping", "PIKE texture generation...", 0.7),
        ("rigging", "Adding facial blendshapes...", 0.8),
        ("animation", "Injecting breathing & micro-motions...", 0.9),
        ("unity_prep", "Preparing for Unity import...", 0.95),
        ("completed", "Ready for 3D world!", 1.0)
    ]
    
    @staticmethod
    async def process_image(job_id: str, image_data: bytes, config: dict):
        """Main processing pipeline"""
        try:
            # Initialize job
            jobs[job_id] = {
                "status": "processing",
                "current_stage": "uploading",
                "progress": 0.1,
                "is_completed": False,
                "error": None
            }
            
            # Step 1: Person Segmentation (MODNet)
            await ProcessingPipeline.segment_person(job_id, image_data)
            
            # Step 2: Depth Estimation (MiDaS/ZoeDepth)
            await ProcessingPipeline.estimate_depth(job_id, image_data)
            
            # Step 3: 3D Mesh Generation (RenderNet)
            await ProcessingPipeline.generate_3d_mesh(job_id)
            
            # Step 4: Texture Mapping (PIKE)
            await ProcessingPipeline.apply_textures(job_id)
            
            # Step 5: Rigging & Animation
            await ProcessingPipeline.add_animations(job_id)
            
            # Step 6: Unity Preparation
            await ProcessingPipeline.prepare_for_unity(job_id)
            
            # Complete
            jobs[job_id].update({
                "current_stage": "completed",
                "progress": 1.0,
                "is_completed": True,
                "assets": {
                    "modelURL": f"https://api.mirrorworld-backend.replit.app/assets/{job_id}/model.glb",
                    "textureURL": f"https://api.mirrorworld-backend.replit.app/assets/{job_id}/texture.jpg",
                    "animationData": f"https://api.mirrorworld-backend.replit.app/assets/{job_id}/animations.json",
                    "metadata": {
                        "facialFeatures": {
                            "eyeColor": "brown",
                            "skinTone": "medium",
                            "facialStructure": {"jawWidth": 0.8, "eyeDistance": 0.6},
                            "expressionCapabilities": ["smile", "blink", "frown"]
                        },
                        "bodyMeasurements": {
                            "height": 1.75,
                            "proportions": {"shoulderWidth": 0.45, "waistWidth": 0.35},
                            "postureData": {"spineAlignment": 0.9}
                        },
                        "animationConfig": {
                            "breathingRate": 16.0,
                            "blinkFrequency": 15.0,
                            "microMotions": {"headSway": 0.3, "weightShift": 0.2},
                            "idlePoses": ["neutral", "slight_lean", "hand_on_hip"]
                        }
                    }
                }
            })
            
        except Exception as e:
            jobs[job_id].update({
                "error": str(e),
                "is_completed": True
            })
    
    @staticmethod
    async def segment_person(job_id: str, image_data: bytes):
        """Step 1: MODNet person segmentation"""
        jobs[job_id].update({
            "current_stage": "segmentation",
            "progress": 0.2
        })
        
        # Simulate MODNet processing
        await asyncio.sleep(2)
        
        # In real implementation:
        # - Load MODNet model
        # - Process image to remove background
        # - Extract person silhouette with high precision
        
        print(f"[{job_id}] Person segmentation completed")
    
    @staticmethod
    async def estimate_depth(job_id: str, image_data: bytes):
        """Step 2: MiDaS/ZoeDepth depth estimation"""
        jobs[job_id].update({
            "current_stage": "depth_analysis",
            "progress": 0.3
        })
        
        await asyncio.sleep(3)
        
        # In real implementation:
        # - Load MiDaS or ZoeDepth model
        # - Generate depth map
        # - Analyze facial geometry and body structure
        
        print(f"[{job_id}] Depth estimation completed")
    
    @staticmethod
    async def generate_3d_mesh(job_id: str):
        """Step 3: RenderNet 3D mesh generation"""
        jobs[job_id].update({
            "current_stage": "mesh_generation",
            "progress": 0.5
        })
        
        await asyncio.sleep(4)
        
        # In real implementation:
        # - Use RenderNet-style neural rendering
        # - Generate high-fidelity 3D mesh
        # - Ensure photorealistic geometry
        
        print(f"[{job_id}] 3D mesh generation completed")
    
    @staticmethod
    async def apply_textures(job_id: str):
        """Step 4: PIKE texture mapping"""
        jobs[job_id].update({
            "current_stage": "texture_mapping",
            "progress": 0.7
        })
        
        await asyncio.sleep(3)
        
        # In real implementation:
        # - Apply PIKE-inspired texture generation
        # - Create photorealistic skin, hair, clothing textures
        # - Add subsurface scattering maps
        
        print(f"[{job_id}] Texture mapping completed")
    
    @staticmethod
    async def add_animations(job_id: str):
        """Step 5: Add facial rigging and animations"""
        jobs[job_id].update({
            "current_stage": "rigging",
            "progress": 0.8
        })
        
        await asyncio.sleep(2)
        
        jobs[job_id].update({
            "current_stage": "animation",
            "progress": 0.9
        })
        
        await asyncio.sleep(2)
        
        # In real implementation:
        # - Add facial blendshapes for expressions
        # - Create breathing animation
        # - Add blinking and micro-motions
        # - Generate idle pose variations
        
        print(f"[{job_id}] Animation rigging completed")
    
    @staticmethod
    async def prepare_for_unity(job_id: str):
        """Step 6: Unity preparation"""
        jobs[job_id].update({
            "current_stage": "unity_prep",
            "progress": 0.95
        })
        
        await asyncio.sleep(1)
        
        # In real implementation:
        # - Export to GLB format with Unity compatibility
        # - Optimize for mobile rendering
        # - Package animations and metadata
        
        print(f"[{job_id}] Unity preparation completed")

@app.post("/api/v1/process-avatar")
async def process_avatar(
    photo: UploadFile = File(...),
    config: str = Form(...)
):
    """Start avatar generation pipeline"""
    try:
        # Generate unique job ID
        job_id = str(uuid.uuid4())
        
        # Read image data
        image_data = await photo.read()
        config_dict = json.loads(config)
        
        # Start processing in background
        asyncio.create_task(ProcessingPipeline.process_image(job_id, image_data, config_dict))
        
        return JSONResponse({
            "jobId": job_id,
            "estimatedTime": 120,  # seconds
            "message": "Avatar generation started"
        })
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/v1/status/{job_id}")
async def get_processing_status(job_id: str):
    """Get processing status for a job"""
    if job_id not in jobs:
        raise HTTPException(status_code=404, detail="Job not found")
    
    job = jobs[job_id]
    return JobStatus(
        job_id=job_id,
        current_stage=job["current_stage"],
        progress=job["progress"],
        is_completed=job["is_completed"],
        assets=job.get("assets"),
        error=job.get("error")
    )

@app.get("/assets/{job_id}/{filename}")
async def get_asset(job_id: str, filename: str):
    """Serve generated assets"""
    # In real implementation, serve actual files
    # For now, return placeholder response
    return JSONResponse({"message": f"Asset {filename} for job {job_id}"})

@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {"status": "healthy", "service": "MirrorWorld AI Backend"}

if __name__ == "__main__":
    # Run on 0.0.0.0 to be accessible from Replit
    uvicorn.run(app, host="0.0.0.0", port=5000)
